Improve host types interface
